7219 move dd from download folders minor refactors
